web: move useVolume component to queries file
- Also adds small code improvements.
